+++ mono_repo/lib/src/github_config.g.dart
@@ -37,8 +37,7 @@ GitHubConfig _$GitHubConfigFromJson(Map json) => $checkedCreate(
           $checkedConvert(
               'on_completion',
               (v) => (v as List<dynamic>?)
-                  ?.map(
-                      (e) => Job.fromJson(Map<String, dynamic>.from(e as Map)))
+                  ?.map((e) => Job.fromJson(e as Map))
                   .toList()),
           $checkedConvert('cron', (v) => v as String?),
           $checkedConvert('stages', (v) => v as List<dynamic>?),