+++ mono_repo/CHANGELOG.md
@@ -2,8 +2,8 @@
 ## 6.4.0-dev
 
 - Added support for `test_with_coverage`.
-  Uses `package:coverage` and [Coveralls](https://coveralls.io) to track test
-  code coverage.
+  Uses `package:coverage` and supports [Coveralls](https://coveralls.io)
+  (the default) and [Codecov](https://codecov.com/) to track test code coverage.
 - Added `--verbose` flag. Helps when debugging failures.
 
 ## 6.3.0